login completo

+++ DAL/Handlers/Models/Response/Response/LoginHandlerResponse.cs
@@ -3,7 +3,7 @@ namespace Amazon.DAL.Handlers.Models.Response.Response
 	public class LoginHandlerResponse
 	{
         public string AccessToken { get; set; }
-		public int IdUser { get; set; }
+		public long IdUser { get; set; }
 		public string Name { get; set; } = String.Empty;
         public string Surname { get; set; } = String.Empty;
